# specs/openapi.yaml
openapi: 3.1.0
info:
  title: bracket-creator API
  description: |
    Web API for generating Kendo tournament brackets.

    This spec covers two servers:
    - **Bracket Generator** (`serve` command): Stateless, form-based bracket generation.
    - **Mobile/Admin API** (`mobile-app` command): Stateful REST API for live tournament management.

    The Mobile/Admin API routes under `/api/` (excluding `/api/viewer/`) require the
    `X-Tournament-Password` header for authentication.
  version: 2.1.0
servers:
  - url: http://localhost:8080
    description: Bracket Generator (serve command)
  - url: http://localhost:8081
    description: Mobile/Admin API (mobile-app command)
...
paths:
...
  /api/competitions/{id}/matches/{mid}/daihyosen:
    post:
      tags: [matches]
      summary: Add representative (daihyosen) bout to a tied team match
      description: >
        Adds a daihyosen (representative) bout to break a tie in an
        elimination team match. Rejected if the match is not tied,
        if it is a pool match, or if neither team has sufficient
        eligible competitors.
      security:
        - TournamentPassword: []
      parameters:
        - $ref: '#/components/parameters/CompetitionId'
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: Daihyosen bout added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResult'
        '400':
          description: Match is not tied (`not_tied`) or is a pool match (`pool_match`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Insufficient eligible competitors (`insufficient_eligibility`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    TournamentPassword:
      type: apiKey
      in: header
      name: X-Tournament-Password
      description: |
        Tournament admin password. Required for all `/api/` routes except
        `/api/viewer/`, `/api/events`, and the bootstrap PUT/POST
        `/api/tournament` while no tournament has been configured yet.

        Header value must exactly equal the stored tournament password
        (exact-string match; no trimming applied to either side).

        Defense-in-depth: if the stored tournament password is somehow
        empty (e.g. operator manually edited `tournament.md` to land
        an empty `password:` field), AuthMiddleware returns 403
        (`tournament misconfigured: password is not set`) for ALL
        requests rather than letting the `header != stored` check
        vacuously pass for empty header.

        Recovery from this state requires **out-of-band repair**: the
        bootstrap PUT/POST `/api/tournament` exception only applies
        when the loaded record is `nil` or matches the literal
        `Name == "New Tournament" && Password == ""` default — a
        real-named record with empty Password is locked out by the 403
        fail-closed branch. Operator must stop the server, edit
        `tournament.md` to either delete it (forces fresh bootstrap)
        or set a non-empty `password:` field in the YAML front-matter,
        then restart.
  parameters:
    CompetitionId:
      in: path
      name: id
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9][a-zA-Z0-9_-]*$'
        maxLength: 64
      description: |
        Competition slug/ID (e.g. "mens-individual"). Must start with an
        alphanumeric character, contain only alphanumerics / hyphens /
        underscores, and be at most 64 chars. Validated server-side via
        `requireValidCompID`; invalid IDs return 400
        (`competition ID contains invalid characters` / `competition ID too long` /
        `competition ID cannot be empty`). The validation is a path-traversal
        defense: without it, an `id` like `../../../etc/passwd` would resolve
        through `filepath.Clean(filepath.Join(...))` and escape the data
        directory.
    MatchId:
      in: path
      name: mid
      required: true
      schema:
        type: string
      description: Match ID (e.g. "pool-a-0" or "bracket-r1-m0").
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    MatchResult:
      type: object
      properties:
        id:
          type: string
        sideA:
          type: string
          maxLength: 100
        sideB:
          type: string
          maxLength: 100
        winner:
          type: string
          maxLength: 100
        ipponsA:
          type: array
          items:
            type: string
          description: "Ippon types scored by side A (M, K, D, T, H)."
        ipponsB:
          type: array
          items:
            type: string
        hansokuA:
          type: integer
        hansokuB:
          type: integer
        decision:
          type: string
          description: >
            Decision code for the match outcome. Extended set:
            `none`/`""` — no decision recorded; `fought` — resolved by scoring;
            `hikiwake` — draw/tie; `kiken` — withdrawal; `fusenpai` — forfeit loss;
            `fusensho` — forfeit win; `daihyosen` — resolved by representative bout;
            `kachinuki-exhaustion` — kachinuki team match ended with one side exhausted.
          enum: [none, fought, hikiwake, kiken, fusenpai, fusensho, daihyosen, kachinuki-exhaustion]
        decisionBy:
          type: string
          description: Which side the decision favours (shiro = white = left, aka = red = right).
          enum: [shiro, aka]
        decisionReason:
          type: string
          maxLength: 200
          description: Free-text operator note for the decision.
        encho:
          $ref: '#/components/schemas/EnchoMetadata'
        status:
          type: string
          enum: [scheduled, running, completed]
        court:
          type: string
          maxLength: 32
        scheduledAt:
          type: string
          maxLength: 32
        subResults:
          type: array
          items:
            $ref: '#/components/schemas/SubMatchResult'
    SubMatchResult:
      type: object
      description: Individual member match result within a team match.
      properties:
        position:
          type: integer
        sideA:
          type: string
          maxLength: 100
        sideB:
          type: string
          maxLength: 100
        ipponsA:
          type: array
          items:
            type: string
        ipponsB:
          type: array
          items:
            type: string
        hansokuA:
          type: integer
        hansokuB:
          type: integer
        winner:
          type: string
          maxLength: 100
        decision:
          type: string
          description: >
            Decision code for this individual sub-match. Same values as `MatchResult.decision`:
            `""` (normal victory) or `"hikiwake"` (draw/tie).
...
    EnchoMetadata:
      type: object
      description: Metadata about extended-time (encho) periods played for a match.
      properties:
        periodCount:
          type: integer
          minimum: 1
...